// frontend/src/api/model/capability.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Apicius API
 * OpenAPI spec version: 1.0.0
 */

export type Capability = typeof Capability[keyof typeof Capability];


export const Capability = {
  BROWSE: 'BROWSE',
  LOOK_UP: 'LOOK_UP',
  ADD: 'ADD',
  UPDATE: 'UPDATE',
  REMOVE: 'REMOVE',
} as const;