// frontend/src/api/model/problemDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Apicius API
 * OpenAPI spec version: 1.0.0
 */
import type { FieldViolation } from './fieldViolation';

export interface ProblemDetail {
  type?: string;
  title?: string;
  status?: number;
  detail?: string;
  violations?: FieldViolation[];
}